Folder facory tweaks
