+++ munch/adventure/Adventure.js
@@ -1,30 +1,12 @@
 const logger = require("../logger.js");
 const path = require("path");
 const { FolderFactory } = require("./FolderFactory.js");
-const { Folder } = require("./Folders/Folder.js");
 const { IdFactory } = require("./IdFactory.js");
 const { TableFactory } = require("./TableFactory.js")
 
 const _ = require("lodash");
 
 class Adventure {
-
-  #createMasterFolders() {
-    const mainRow = { id: -1, cobaltId: -1, title: this.config.run.book.description };
-
-    const folderData = {
-      adventure: this,
-      row: mainRow,
-      specialType: "base",
-    };
-
-    this.masterFolder = {
-      JournalEntry: new Folder(_.merge(folderData, { type: "JournalEntry"})).toJson(),
-      Scene: new Folder(_.merge(folderData, { type: "Scene"})).toJson(),
-      RollTable: new Folder(_.merge(folderData, { type: "RollTable"})).toJson(),
-      Actor: new Folder(_.merge(folderData, { type: "Actor"})).toJson(),
-    };
-  }
 
   constructor(config) {
     this.config = config;
@@ -61,19 +43,25 @@ class Adventure {
 
     this.masterFolder = {};
 
-    // create factories
+    // create global factories
     this.idFactory = new IdFactory(this);
     this.folderFactory = new FolderFactory(this);
-    this.journalFactory = new JournalFactory(this);
-    this.tableFactory = new TableFactory(this);
+    
+    
+    
+    // These factories are per document and rendered per html doc.
+    // this.tableFactory = new TableFactory(this);
+    // this.journalFactory = new JournalFactory(this);
+
 
     // initialize master folders
-    this.#createMasterFolders();
+    this.masterFolder = this.folderFactory.masterFolders;
 
   }
 
-  // check to see if I need to override the defaults
   toJson() {
+    // TODO:
+    // loop through attached arrays and render out json objects
     // return JSON.stringify(this.data);
   }
